Allows some recovery addresses to override others.
This is achieved by having each recovery address have a number of recovery days, rather than hard-coding it for the whole wallet.  An address with a shorter recovery period can override a recovery undertaken by an address with a longer recovery period.

The purpose of this change is to allow someone to have a tiered recovery system such as a hardware wallet with short recovery time that can override other recovery attempts, and a less secure final fallback such as drop-box mnemonic which has a long recovery time and can be overriden by other recovery addresses.