Rewrite using TypeScript; Add support for proxy calls.
While I very much prefer Kotlin over TypeScript, the Ethereum ecosystem is pretty soundly seated in JS land and continuing to work in Kotlin/Java land is not worth the effort at this time.  This commit migrates all tooling and infrastructure over to TypeScript and also includes major updates to the wallet itself.