Moves the browser extension/test stuff over to its own project.
https://github.com/Zoltu/ethereum-browser-sdk

Normally I would rebase all of this out of this branch, but the OpenZeppelin audit was tagged on specific commits which I need to ensure remain in history, so this is as far back as I can go to delete this stuff without causing commit hashes to change.  You should have seen how much more code I wrote beyond what is here that I _was_ able to excise from history!