Updates readme.
Fixes toolbox and updates a bunch of toolbox dependencies in order to ensure the readme example code is correct-ish.

Updates `deployment-parameters`, as it was not updated the last time the solidity file was touched.  It only contained a comment change, but it feels  the source code should match the deployment code exactly.