Renames `addOrUpdate` to `set` and fixes contract bug.
I believe this results in a simpler mental model, where you have `get` (which gets the current value for the given key, or throws), `set` (which sets the current value for the given key), and `remove` (which removes the current key).

Bug was that trying to remove the key in the last position would fail as it was reading off the end of the array in that case.